API/requirements.txt
flask
uvicorn
sqlalchemy
pymysql
mysql-connector-python 